rearrange coroutine core code and flesh out its documentation
