+++ kernel/src/coroutine/mod.rs
@@ -262,94 +262,6 @@
 //! A connector may drop any [`Resume`] handle at any time. This abandons the suspended coroutine
 //! and drops all associated state. For connector-side work failures, resume with `Err` when kernel
 //! should observe the failure; otherwise drop the handle and return the error directly.
-// === Kernel implementation ===
-//
-// Kernel workflows are ordinary async Rust functions, but this module allows the connector to drive
-// them directly, without runtime scheduling: Kernel futures are polled with a no-op waker and
-// advance only on explicit start or resume calls:
-//
-// 1. The connector calls a kernel entry point that creates a `Workflow`.
-//
-// 2. The entry point calls `Workflow::start` with a closure that, when invoked synchronously,
-//    associates the workflow's async logic with a `Channel` and returns the resulting async future.
-//
-// 3. `Workflow::start` creates an `Arc<Mailbox<PendingRequest>>`. It keeps one reference for
-//    itself, and shares a second reference with kernel by wrapping it in a `Channel`, which it uses
-//    to invoke the workflow-creation closure. Because that closure returns a future, invoking it
-//    creates but does not start the workflow's (compiler-generated) async state machine.
-//
-// 4. `Workflow::start` wraps and instruments that future, and box-pins it (type alias: `Task<O>`).
-//    This `Task` owns the complete compiler-generated state of the kernel workflow.
-//
-// 5. `Workflow::start` passes the task and mailbox to `advance_workflow`.
-//
-// 6. `advance_workflow` creates a `Context` with a no-op waker and explicitly polls the task. No
-//    async runtime participates in coroutine scheduling.
-//
-// 7. Polling the task launches it (standard async Rust), executing the kernel workflow until it
-//    calls an async channel method such as `channel.read_small_file(...)`. All channel methods are
-//    thin wrappers that call and await the async `Channel::exchange` method.
-//
-// 8. The kernel’s `.await` polls that channel-method future. Its body calls
-//    `Channel::exchange(...)`, creating and immediately awaiting another compiler-generated future.
-//
-// 9. The first poll of the `exchange` future creates an `Arc<Exchange>` in `Outbound(operation)`
-//    state and stores a `Weak<Exchange>` in the channel mailbox. The weak pointer will be dropped
-//    automatically if kernel abandons the exchange after publishing it to the connector.
-//
-// 10. The `exchange` future then constructs `Wait(exchange)` and awaits it. The generated
-//     `exchange` future invokes `Wait::poll`.
-//
-// 11. `Wait::poll` sees `Outbound`, leaves that state intact, and returns `Pending`.
-//
-// 12. `Pending` propagates through the `exchange` future, channel-method future, and workflow
-//     future. Their compiler-generated async state machines retain the nested futures and all
-//     suspended locals, including the only strong reference to the `Exchange`. The task’s poll
-//     finally returns `Pending` to `advance_workflow`.
-//
-// 13. After that poll returns, `advance_workflow` removes the pending request from the mailbox and
-//     calls `PendingRequest::attach`.
-//
-// 14. `attach` upgrades the mailbox’s `Weak<Exchange>`, claims the operation, and transitions the
-//     exchange from `Outbound` to `InFlight`.
-//
-// 15. `attach` creates a `TypedResume` whose closure owns the upgraded `Arc<Exchange>`, pinned
-//     workflow task (which still owns the original `Arc<Exchange>`), mailbox, and the function
-//     needed to advance that task again. It combines the claimed operation with this resume handle
-//     to produce a typed `Request`.
-//
-// 16. `advance_workflow` returns `Workflow::Request(request)` to `Workflow::start`, which returns
-//     it to the connector. At this point, the `Mailbox`, `Channel`, and `Exchange` are all
-//     reachable only through the `TypedResume`. If the connector drops that `request`, the entire
-//     workflow and all associated state are dropped as Arc reference counts go to zero.
-//
-// 17. The connector matches the request, performs the operation, and calls
-//     `TypedResume::resume(response)`. The resume closure stores the response in the exchange,
-//     transitioning `InFlight` to `Inbound(response)`, then forwards the `Task` and the `Mailbox`
-//     to `advance_workflow`.
-//
-// 18. `advance_workflow` polls the task again, and the compiler-generated async state machines
-//     resume the nested poll chain at the suspended `.await`. The `Wait` instance now sees
-//     `Inbound(response)`, transitions the exchange to `Complete`, and returns `Ready(response)`.
-//     That result propagates through the `exchange` and channel-method futures, and becomes the
-//     return value of kernel’s original `.await`.
-//
-// 19. The kernel workflow continues until it creates and awaits another channel future, repeating
-//     the process, or returns its final `DeltaResult<O>`. `advance_workflow` converts successful
-//     completion to `Workflow::Done(output)` and propagates any error directly.
-//
-// `Yielder` adds `yield_item`, which logically separates the entity that consumes yielded items
-// from the entity that processes requests. When kernel drives a child generator, it consumes the
-// yielded items directly while forwarding its requests back to the connector through its own
-// `Channel`. The connector therefore sees those requests as part of the overall workflow's
-// coroutine rather than driving a separate child generator. Kernel uses these internal generators
-// to manipulate streaming outputs produced by e.g. log replay.
-//
-// For connectors that don't want to deal with coroutines at all, kernel continues to provide legacy
-// sync/imperative entry points. Each entry point launches the corresponding workflow, then drives
-// it to completion using an `EngineConnector` that uses the caller-provided `Engine` instance to
-// serve requestse.
-
 use std::any::Any;
 use std::marker::PhantomData;
